test: move the shared test code into its own module
The test code must reside in its own module, then get compiled in using testImplementation and
androidTestImplementation as the sourcesets approach doesn't work across module; it will show the
files existence, but you cannot import it
