+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':app', ':account', ':database', ':core', ':series', ':app-discover', ':app-timeline', ':app-profile',
+include ':app', ':account', ':database', ':core', ':series', ':app-discover', ':app-timeline', ':app-profile', ':testing',
         ':server',
         ':kitsu'
